requirements.txt
Flask==2.3.2
ultralytics==8.0.196
opencv-python==4.9.0.80  # Updated to the correct version
numpy==1.24.3
tensorflow==2.15.0
keras==2.15.0
Pillow==9.5.0
python-dotenv==1.0.1
gunicorn==22.0.0